Add files via upload
These are the different from my thesis. the "dressel" files are my recreation of the experiments done by dressel and farid as detailed in their paper titles "The accuracy, fairness, and limits "of predicting recidivism. The modified, bigdata, and split race files are all variation of the dressel model that I made to make a comparison.